// web/apps/msteams/src/components/booking/editBooking/__generated__/editBooking_organizationMembers_query.graphql.ts
export type editBooking_organizationMembers_query$data = {
  readonly organizationMembers: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly customer: {
          readonly familyName: string | null | undefined;
          readonly givenName: string | null | undefined;
          readonly middleName: string | null | undefined;
          readonly name: string | null | undefined;
          readonly photoUrl: string | null | undefined;
          readonly uniqueId: string;
        };
        readonly id: string;
      };
    }>;
    readonly totalCount: number | null | undefined;
  } | null | undefined;
  readonly " $fragmentType": "editBooking_organizationMembers_query";
};
export type editBooking_organizationMembers_query$key = {
  readonly " $data"?: editBooking_organizationMembers_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"editBooking_organizationMembers_query">;
};